updated withdral

+++ wallet/views.py
@@ -1,15 +1,17 @@
-from django.shortcuts import get_object_or_404, render
+from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
-from wallet.models import CryptoAmount, Wallet
+from flask import redirect
+from api_backend.models import CryptoCurrency
+from wallet.models import CryptoWallet, CryptoCollection
 
 @login_required(login_url="account_login")
 def wallet_view(request):
     template_name = "wallet/wallet.html"
     user = request.user
-    wallet, _ = Wallet.objects.get_or_create(owner=user)
+    wallet, _ = CryptoWallet.objects.get_or_create(owner=user)
     crypto_data = []
 
-    for crypto_amount in CryptoAmount.objects.filter(wallet=wallet):
+    for crypto_amount in CryptoCollection.objects.filter(wallet=wallet):
         crypto = crypto_amount.cryptocurrency
         amount = crypto_amount.amount
 
@@ -30,10 +32,13 @@ def wallet_view(request):
 
 
 @login_required(login_url="account_login")
-def withdraw(request, pk):
-    template_name = "wallet/wallet.html"
+def withdrawal(request, pk):
     user = request.user
-    wallet = get_object_or_404(Wallet, owner=user)
-    crypto_amount = get_object_or_404(CryptoAmount, pk=pk, wallet=wallet)
-    crypto_amount.delete()
-    return render(request, template_name, {'wallet': wallet})
+    wallet = CryptoWallet.objects.filter(owner=user).first()
+    crypto = CryptoCurrency.objects.filter(id=pk).first()
+    crypto_amounts = CryptoCollection.objects.filter(wallet=wallet)
+    for crypto_amount in crypto_amounts:
+        if crypto_amount.cryptocurrency == crypto:
+            crypto_amount.delete()
+
+    return redirect("wallet")
